refactor(web): extract WorkoutCalendarBoard from Calendar.tsx
Lifts the visible-month state, workouts cache, grid render, and the
WorkoutDrawer mount into a reusable `<WorkoutCalendarBoard>` so the
upcoming `/personal-program` page (#183) mounts the exact same calendar
+ prescription drawer the gym `/calendar` uses, with no fork.

Pure refactor — gym `/calendar` behavior is identical:
- 193/193 web vitest suite passes
- E2E `tests/workout-publish.spec.ts` passes (PROGRAMMER creates a draft
  via the calendar drawer + MEMBER feed shows the published workout)

Shape:
- New `WorkoutCalendarBoard` props: `loadWorkouts(from, to)` callback
  (caller binds the per-gym filtered list call, since `ProgramScope.listWorkouts`
  returns *all* workouts in a single program — date-range paging across
  multiple programs needs a separate hook), plus the existing
  `scope` / `userGymRole` / `defaultProgramId` from #197 forwarded to
  `WorkoutDrawer`.
- `Calendar.tsx` becomes a thin wrapper: builds the gym scope + `loadWorkouts`
  closure (api.workouts.list with movement + program filters) and renders
  its existing program-filter / movement-filter chrome above the board.

Part of #183
