+++ apps/web/src/pages/Calendar.tsx
@@ -1,24 +1,13 @@
-import { useState, useEffect, useCallback, useMemo } from 'react'
+import { useCallback, useMemo, useState } from 'react'
 import { Link } from 'react-router-dom'
-import { api, type Workout } from '../lib/api'
+import { api } from '../lib/api'
 import { useGym } from '../context/GymContext.tsx'
 import { makeGymProgramScope } from '../lib/gymProgramScope'
 import { useMovements } from '../context/MovementsContext.tsx'
 import { useProgramFilter } from '../context/ProgramFilterContext.tsx'
-import CalendarCell from '../components/CalendarCell'
-import WorkoutDrawer from '../components/WorkoutDrawer'
+import WorkoutCalendarBoard from '../components/WorkoutCalendarBoard'
 import MovementFilterInput from '../components/MovementFilterInput'
-import Button from '../components/ui/Button'
 import Chip from '../components/ui/Chip'
-
-function toDateKey(date: Date): string {
-  const y = date.getFullYear()
-  const m = String(date.getMonth() + 1).padStart(2, '0')
-  const d = String(date.getDate()).padStart(2, '0')
-  return `${y}-${m}-${d}`
-}
-
-const DAY_HEADERS = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat']
 
 export default function Calendar() {
   const { gymId, gymRole: userGymRole } = useGym()
@@ -28,46 +17,9 @@ export default function Calendar() {
   )
   const allMovements = useMovements()
   const { selected: programIds, available, clear: clearProgramFilter } = useProgramFilter()
-  const today = new Date()
-  const [year, setYear] = useState(today.getFullYear())
-  const [month, setMonth] = useState(today.getMonth())
-  const [workouts, setWorkouts] = useState<Workout[]>([])
-  const [loading, setLoading] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-  const [selectedDate, setSelectedDate] = useState<string | null>(null)
-  const [selectedWorkoutId, setSelectedWorkoutId] = useState<string | null>(null)
   const [filterMovementIds, setFilterMovementIds] = useState<string[]>([])
 
   const programIdsKey = programIds.join(',')
-
-  const loadWorkouts = useCallback(async (signal?: { cancelled: boolean }) => {
-    if (!gymId) return
-    setLoading(true)
-    setError(null)
-    try {
-      const from = new Date(year, month, 1).toISOString()
-      const to = new Date(year, month + 1, 0, 23, 59, 59, 999).toISOString()
-      const filters = (filterMovementIds.length || programIds.length)
-        ? {
-            ...(filterMovementIds.length ? { movementIds: filterMovementIds } : {}),
-            ...(programIds.length ? { programIds } : {}),
-          }
-        : undefined
-      const data = await api.workouts.list(gymId, from, to, filters)
-      if (!signal?.cancelled) setWorkouts(data)
-    } catch (e) {
-      if (!signal?.cancelled) setError((e as Error).message)
-    } finally {
-      if (!signal?.cancelled) setLoading(false)
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [gymId, year, month, filterMovementIds, programIdsKey])
-
-  useEffect(() => {
-    const signal = { cancelled: false }
-    loadWorkouts(signal)
-    return () => { signal.cancelled = true }
-  }, [loadWorkouts])
 
   // Single-program filter gets a featured header (color stripe + name).
   // Multi-program gets a compact "Filtered to N programs" eyebrow.
@@ -79,47 +31,17 @@ export default function Calendar() {
   // the drawer's existing "first program in list" behavior beyond N=1.
   const defaultProgramIdForCreate = programIds.length === 1 ? programIds[0] : undefined
 
-  const workoutsByDate: Record<string, Workout[]> = {}
-  for (const w of workouts) {
-    // scheduledAt is UTC midnight — slice the ISO string to get the UTC calendar date,
-    // avoiding a local-timezone shift that would bucket the workout a day early for US users.
-    const key = w.scheduledAt.slice(0, 10)
-    if (!workoutsByDate[key]) workoutsByDate[key] = []
-    workoutsByDate[key].push(w)
-  }
-
-  const workoutsOnDay = selectedDate ? (workoutsByDate[selectedDate] ?? []) : []
-  const selectedWorkout = selectedWorkoutId
-    ? workoutsOnDay.find(w => w.id === selectedWorkoutId)
-    : undefined
-
-  // Build padded grid of Date | null
-  const firstDayOfMonth = new Date(year, month, 1)
-  const daysInMonth = new Date(year, month + 1, 0).getDate()
-  const startDow = firstDayOfMonth.getDay()
-  const cells: (Date | null)[] = []
-  for (let i = 0; i < startDow; i++) cells.push(null)
-  for (let d = 1; d <= daysInMonth; d++) cells.push(new Date(year, month, d))
-  while (cells.length % 7 !== 0) cells.push(null)
-  const weeks: (Date | null)[][] = []
-  for (let i = 0; i < cells.length; i += 7) weeks.push(cells.slice(i, i + 7))
-
-  const monthLabel = firstDayOfMonth.toLocaleString('default', { month: 'long', year: 'numeric' })
-  const todayKey = toDateKey(today)
-
-  function prevMonth() {
-    if (month === 0) { setYear(y => y - 1); setMonth(11) }
-    else setMonth(m => m - 1)
-    setSelectedDate(null)
-    setSelectedWorkoutId(null)
-  }
-
-  function nextMonth() {
-    if (month === 11) { setYear(y => y + 1); setMonth(0) }
-    else setMonth(m => m + 1)
-    setSelectedDate(null)
-    setSelectedWorkoutId(null)
-  }
+  const loadWorkouts = useCallback(async (from: string, to: string) => {
+    if (!gymId) return []
+    const filters = (filterMovementIds.length || programIds.length)
+      ? {
+          ...(filterMovementIds.length ? { movementIds: filterMovementIds } : {}),
+          ...(programIds.length ? { programIds } : {}),
+        }
+      : undefined
+    return api.workouts.list(gymId, from, to, filters)
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [gymId, filterMovementIds, programIdsKey])
 
   if (!gymId) {
     return (
@@ -145,7 +67,7 @@ export default function Calendar() {
       )}
 
       {/* Header */}
-      <div className="flex items-center justify-between mb-6">
+      <div className="mb-6">
         {singleProgram ? (
           <div className="flex items-start gap-3 min-w-0">
             <div
@@ -167,15 +89,6 @@ export default function Calendar() {
         ) : (
           <h1 className="text-2xl font-bold">Calendar</h1>
         )}
-        <div className="flex items-center gap-2">
-          <Button variant="tertiary" onClick={prevMonth} aria-label="Previous month">
-            ←
-          </Button>
-          <span className="text-base font-medium w-44 text-center select-none">{monthLabel}</span>
-          <Button variant="tertiary" onClick={nextMonth} aria-label="Next month">
-            →
-          </Button>
-        </div>
       </div>
 
       {/* Sticky movement-filter sub-header */}
@@ -207,58 +120,11 @@ export default function Calendar() {
         </div>
       )}
 
-      {error && <p className="text-red-400 mb-4">{error}</p>}
-
-      {/* Day-of-week headers */}
-      <div className="grid grid-cols-7 mb-px">
-        {DAY_HEADERS.map((d) => (
-          <div key={d} className="text-center text-xs text-gray-400 py-1">
-            {d}
-          </div>
-        ))}
-      </div>
-
-      {/* Calendar grid */}
-      <div
-        className={[
-          'grid grid-cols-7 gap-px bg-gray-800 border border-gray-800 rounded-lg overflow-hidden',
-          loading ? 'opacity-60 pointer-events-none' : '',
-        ].join(' ')}
-      >
-        {weeks.map((week, wi) =>
-          week.map((date, di) => {
-            if (!date) {
-              return <div key={`empty-${wi}-${di}`} className="bg-gray-950 h-[128px]" />
-            }
-            const key = toDateKey(date)
-            return (
-              <CalendarCell
-                key={key}
-                date={date}
-                isToday={key === todayKey}
-                workouts={workoutsByDate[key] ?? []}
-                selected={key === selectedDate}
-                onAddClick={() => { setSelectedDate(key); setSelectedWorkoutId(null) }}
-                onWorkoutClick={(id) => { setSelectedDate(key); setSelectedWorkoutId(id) }}
-              />
-            )
-          }),
-        )}
-      </div>
-
-      <WorkoutDrawer
+      <WorkoutCalendarBoard
+        loadWorkouts={loadWorkouts}
         scope={scope}
-        dateKey={selectedDate}
-        workout={selectedWorkout}
-        workoutsOnDay={workoutsOnDay}
         userGymRole={userGymRole}
         defaultProgramId={defaultProgramIdForCreate}
-        onClose={() => { setSelectedDate(null); setSelectedWorkoutId(null) }}
-        onSaved={() => { setSelectedDate(null); setSelectedWorkoutId(null); loadWorkouts() }}
-        onAutoSaved={loadWorkouts}
-        onReordered={loadWorkouts}
-        onWorkoutSelect={(id) => setSelectedWorkoutId(id)}
-        onNewWorkout={() => setSelectedWorkoutId(null)}
       />
     </div>
   )
